% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_Etfs_List.R
\name{EN_Etfs_List}
\alias{EN_Etfs_List}
\title{Get the list of Etfs quoted on Euronext markets}
\usage{
EN_Etfs_List()
}
\value{
A data frame containing information about Etfs, including name, ticker, ISIN code, market, last price, percentage change, and URL.
}
\description{
This function retrieves information about Etfs quoted on Euronext markets.
No input parameters are required for this function. It retrieves information about Etfs quoted on Euronext markets
}
\examples{

\dontrun{

dt_ = EN_Etfs_List()
print(dt_)
}

}
\seealso{
\code{\link{EN_Stocks_List}} to get the list tickers infos quoted on a given Euronext Stock,
\code{\link{EN_Bonds_List}} for Bonds list,  \code{\link{EN_Indices_List}} for Indexes list,
\code{\link{EN_GetISIN_Etf}}  to get the ADN(Isin and Market indentifier) of a givin Etf.

Other Infos Retrieval: 
\code{\link{EN_Bonds_List}()},
\code{\link{EN_Funds_List}()},
\code{\link{EN_GetISIN_B}()},
\code{\link{EN_GetISIN_Etf}()},
\code{\link{EN_GetISIN_F}()},
\code{\link{EN_Get_News}()},
\code{\link{EN_Indices_List}()},
\code{\link{EN_Stocks_List}()}
}
\concept{Infos Retrieval}
