% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_Funds_List.R
\name{EN_Funds_List}
\alias{EN_Funds_List}
\title{Get the list of funds quoted on Euronext markets}
\usage{
EN_Funds_List()
}
\value{
A data frame containing information about funds, including name, ticker, ISIN code, market, last price, percentage change, and URL.
}
\description{
This function retrieves information about funds quoted on Euronext markets.
No input parameters are required for this function. It retrieves information about Funds quoted on Euronext markets
}
\examples{
\donttest{

library(httr)
library(jsonlite)
library(rvest)
library(stringr)

dt_ = EN_Funds_List()
print(dt_)
}

}
\seealso{
\code{\link{EN_Stocks_List}} to get the list tickers infos quoted on a given Euronext Stock,
\code{\link{EN_Bonds_List}} for Bonds list,  \code{\link{EN_Indices_List}} for Indexes list,
\code{\link{EN_Etfs_List}}  for Etfs list

Other Infos Retrieval: 
\code{\link{EN_Bonds_List_bis}()},
\code{\link{EN_Bonds_List}()},
\code{\link{EN_Etfs_List_bis}()},
\code{\link{EN_Etfs_List}()},
\code{\link{EN_GetISIN_B}()},
\code{\link{EN_GetISIN_Etf}()},
\code{\link{EN_GetISIN_F}()},
\code{\link{EN_Get_News}()},
\code{\link{EN_Indices_List}()},
\code{\link{EN_Stocks_List}()},
\code{\link{En_Market.Sumarry}()}
}
\concept{Infos Retrieval}
