% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_GetISIN.R
\name{EN_GetISIN}
\alias{EN_GetISIN}
\title{Get ISIN of a company listed on Euronext}
\usage{
EN_GetISIN(ticker)
}
\arguments{
\item{ticker}{A character string representing the company's ticker, name or ISIN.}
}
\value{
A character string representing the ISIN of the company.
}
\description{
This function retrieves the ISIN of a company listed on Euronext Paris
using the provided ticker. The function formats the ticker to uppercase
and makes an API request to Euronext to get the corresponding ISIN.
}
\examples{
library(httr)
# Get ISIN for the ticker "4DDD"
EN_GetISIN("4DDD")

# Get ISIN for the ticker "ADS MARITIME HOLD"
EN_GetISIN("ADS MARITIME HOLD")

# Get ISIN for ATENOR company
EN_GetISIN("BE0003837540")

}
