% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_GetISIN_bis.R
\name{EN_GetISIN_bis}
\alias{EN_GetISIN_bis}
\title{Get a caracteristic (Name, ISIN, DNA, Symbol) of a Stock or an Index listed on Euronext}
\usage{
EN_GetISIN_bis(ticker)
}
\arguments{
\item{ticker}{A character string representing the Stock's or an Index's ticker, name, or ISIN.}
}
\value{
A List (Name, ISIN, DNA, Symbol) representing the DNA of the Stock or an Index.
}
\description{
This function retrieves the caracteristic elements of a Stock or an Index listed on Euronext
using the provided Symbol/Name/ISIN of a Stock or an Index. The function formats the ticker to uppercase
and makes an API request to Euronext to get the corresponding caracteristics.
}
\examples{

\dontrun{
library(httr)
# Get caracteristics for the ticker "4DDD"
EN_GetISIN_bis("4DDD")

# Get caracteristics for the ticker "ADS MARITIME HOLD"
EN_GetISIN_bis("ADS MARITIME HOLD")

# Get caracteristics for ATENOR company
EN_GetISIN_bis("BE0003837540")

# Get caracteristics for AEX All-Share Index GR
EN_GetISIN_bis("AEX All-Share Index GR")

# Get caracteristics of AEX Health Care Index
EN_GetISIN_bis("NLHC")
}

}
