% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_GetProfile.R
\name{EN_GetProfile}
\alias{EN_GetProfile}
\title{Get a caracteristic (list of Name, ISIN, DNA, Symbol/or Issuer) of a Stock, Index, Etfs, Fund or Bond listed on Euronext}
\usage{
EN_GetProfile(ticker, stock_type = "Eq_Ind")
}
\arguments{
\item{ticker}{A character string representing the Stock's or an Index's ticker, name, or ISIN.}

\item{stock_type}{The type of the ticker: 'Eq_Ind' for Stocks and Indexes, 'Fund' or "F" for Fund tickers, 'Bond' or "B" for Bond tickers, and 'Etfs' or "E" for EFTs.}
}
\value{
A List (Name, ISIN, DNA, Symbol) representing the DNA of the Stock or an Index.
}
\description{
This function retrieves the caracteristic elements of a a Stock, Index, Etfs, Fund or Bond listed on Euronext
using the provided Symbol/Name/ISIN or DNA.
}
\examples{

\dontrun{
library(httr)
# Examples for Equities and Indexes
# Note : Provide the equity Symbol, ISIN, or Name for best results
# Example: '3D SYSTEMS CORP' equity
EN_GetProfile("4ddd")  # Provide Symbol

# Example: 'ALFEN' equity
EN_GetProfile("NL0012817175") # Provide ISIN

# Example: 'LES HOTELS BAVEREZ' equity
EN_GetProfile("LES HOTELS BAVEREZ") # Provide Name

# Example for ETFs
# Provide the equity Symbol, ISIN, Name, or DNA for best results
# Example: Asia IG Corp US A
EN_GetProfile("$Asia IG Corp US A", stock_type = 'E') # Provide Name
# Or
EN_GetProfile("MSFT", stock_type = 'E') # Provide Symbol
# Or
EN_GetProfile("XS2399367254", stock_type = 'E') # Provide ISIN of 3X LONG COINBASE
# Or
EN_GetProfile("IE0007G78AC4-XAMS", stock_type = 'E') # Provide DNA

# Example for Funds
# Note : Provide the equity Symbol, ISIN, Name, or DNA for best results
# Example: ACOMEA PERFORMANCE
EN_GetProfile("ACAPER", stock_type = 'F') # Provide Symbol
# Or
EN_GetProfile("BNP ESGNL", stock_type = 'F') # Provide Name
# Or
EN_GetProfile("NL0015000W40", stock_type = 'F') # Provide ISIN of SWIF2
# Or
EN_GetProfile("NL0000293181-XAMS", stock_type = 'F') # Provide DNA of GSDM5

# Examples for Bonds
# Note : Provide the Bond DNA, and sometimes its Name for best results
# Example: 'A2A SLB TF 0,625\%' Bond
EN_GetProfile("XS2364001078-XMOT", stock_type = 'B') # Provide DNA

# Example: 'AAB1.50\%30SEP30' Bond
EN_GetProfile("AAB1.50\%30SEP30", stock_type = 'B') # Provide Name

}

}
