% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_Ticker_Perfomance.R
\name{EN_Ticker_Perfomance}
\alias{EN_Ticker_Perfomance}
\title{Get Ticker Historical Performance}
\usage{
EN_Ticker_Perfomance(ticker, escape = FALSE, stock_type = "Eq_Ind")
}
\arguments{
\item{ticker}{A character string representing the company's ticker symbol.}

\item{escape}{Boolean, either T or F. If escape is True, it means you're providing the DNA (ISIN-Market identifier) directly. Giving T to escape is helpful to avoid time-consuming operations; otherwise, F means you need to provide the Ticker symbol, name, or ISIN and the type of market to which it belongs.}

\item{stock_type}{The type of the ticker: 'Eq_Ind' for Stocks and Indexes, 'Fund' or "F" for Fund tickers, 'Bond' or "B" for Bond tickers, and 'Etfs' or "E" for EFTs.}
}
\value{
A data frame containing historical performance data of the specified company.
If the ticker is not found, it returns the message "Ticker not found."
}
\description{
This function retrieves the historical performance data of a company listed on the Euronext exchange
based on its ticker symbol. It returns a table with details such as the highest and lowest stock prices,
date of the highest and lowest prices, and other relevant information.
}
\examples{
\dontrun{
library(httr)
library(httr2)
library(rvest)
library(jsonlite)
library(stringr)

Get Performance of Ticker ABCA
dt = EN_Ticker_Perfomance("ABCA")
print(dt)
# Get Performance of ETF AAPL
EN_Ticker_Perfomance("AAPL", stock_type = "E", escape = FALSE)

# Get Performance of Bond issued by A2A S.p.A.
dt1 = EN_Ticker_Perfomance("XS1195347478-ETLX", escape = TRUE)
print(dt1)

# Get Performance of ACOMEA EMERGING Q2 Fund
dt2 = EN_Ticker_Perfomance("IT0005091126", 'F', escape = FALSE)
print(dt2)

# Get Performance of AEX All-Share Index
dt3 = EN_Ticker_Perfomance("AEX All-Share Index GR", escape = FALSE)
print(dt3)

}

}
\seealso{
\code{\link{EN_GetISIN}} to retrieve the ISIN for a given ticker.

Other Data Retrieval: 
\code{\link{EN_Get_OB}()},
\code{\link{EN_HistData_bis}()},
\code{\link{EN_HistData}()},
\code{\link{EN_Ticker_hcData}()},
\code{\link{EN_Ticker_infos}()},
\code{\link{EN_intraday_Data}()}

Other Euronext: 
\code{\link{EN_Get_News}()},
\code{\link{EN_Get_OB}()},
\code{\link{EN_HistData_bis}()},
\code{\link{EN_HistData}()},
\code{\link{EN_P_Ticker_adn}()},
\code{\link{EN_Ticker_hcData}()},
\code{\link{EN_Ticker_infos}()},
\code{\link{EN_intraday_Data}()}
}
\concept{Data Retrieval}
\concept{Euronext}
