% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_Ticker_hcData.R
\name{EN_Ticker_hcData}
\alias{EN_Ticker_hcData}
\title{Get Highchart Data: Price and Volume}
\usage{
EN_Ticker_hcData(
  ticker,
  escape = F,
  stock_type = "Eq_Ind",
  from = NULL,
  to = Sys.Date()
)
}
\arguments{
\item{ticker}{A character string representing the company's ticker, name, or ISIN.}

\item{escape}{Boolean, either T or F. If escape is True, it means you're providing the DNA (ISIN-Market identifier) directly. Giving T to escape is helpful to avoid time-consuming operations; otherwise, F means you need to provide the Ticker symbol, name, or ISIN and the type of market to which it belongs.}

\item{stock_type}{The type of the ticker: 'Eq_Ind' for Stocks and Indexes, 'Fund' or "F" for Fund tickers, 'Bond' or "B" for Bond tickers, and 'Etfs' or "E" for EFTs.}

\item{from}{A Date object specifying the start date for the data. Default is \code{NULL}, which retrieves data from the earliest available date. Users can adjust the value of \code{from} to specify a different start date. Only dates in the format "\%Y-\%m-\%d" are allowed.}

\item{to}{A Date object specifying the end date for the data. Default is the current date. Users can adjust the value of \code{to} to specify a different end date. Only dates in the format "\%Y-\%m-\%d" are allowed.}
}
\value{
A tibble containing historical price and volume data for the specified stock ticker within the specified date range.
If the specified ticker is not found, the function returns an error message.
}
\description{
This function retrieves historical price and volume data for a given stock ticker on the Euronext exchange.
The data is returned as a tibble with columns for Date, Price, and Volume. You can specify the date range
for which you want to retrieve the data using the 'from' and 'to' parameters.
}
\examples{
\dontrun{
library(dplyr)
# Get company ABC ARBITRAGE share hc data
dt_ABCA = EN_Ticker_hcData("ABca")
print(dt_ABCA)

dt_EOS = EN_Ticker_hcData("EOS", from = "2000-10-30", to = "2024-01-27")
print(dt_EOS)

# Get hc data for ETF AAPL from December 15, 2023 to now
hc_aapl <- EN_Ticker_hcData("AAPL", stock_type = "E", from = "2023-12-15", to = Sys.Date())
head(hc_aapl)

# Get hc data of Fund KEMPEN GBL HDiv N
dt_kem = EN_Ticker_hcData("US88554D2053-ETLX", escape = T)
print(dt_kem)

#' #To get hc data of Bond issued by A2A S.p.A.
hc_RMBS <- EN_Ticker_hcData("XS1195347478-ETLX", escape = T)
head(hc_RMBS)

}

}
\seealso{
\code{\link{EN_GetISIN}} to retrieve the ISIN for a given ticker,
\code{\link{EN_HistData}} to retrieve daily data for a given ticker.

Other Data Retrieval: 
\code{\link{EN_Get_OB}()},
\code{\link{EN_HistData}()},
\code{\link{EN_Ticker_Perfomance}()},
\code{\link{EN_Ticker_infos}()},
\code{\link{EN_intraday_Data}()}

Other Euronext: 
\code{\link{EN_Get_News}()},
\code{\link{EN_Get_OB}()},
\code{\link{EN_HistData}()},
\code{\link{EN_P_Ticker_adn}()},
\code{\link{EN_Ticker_Perfomance}()},
\code{\link{EN_Ticker_infos}()},
\code{\link{EN_intraday_Data}()}
}
\concept{Data Retrieval}
\concept{Euronext}
